// frontend/fingertips-frontend/generated-sources/ft-api-client/models/ReportingPeriod.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Fingertips API
 * An API to query public health indicator data.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: ProfileFeedback@dhsc.gov.uk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * The time period that a single data point represents.
 * @export
 */
export const ReportingPeriod = {
    Monthly: 'Monthly',
    Quarterly: 'Quarterly',
    CumulativeQuarterly: 'CumulativeQuarterly',
    Yearly: 'Yearly',
    TwoYearly: 'TwoYearly',
    ThreeYearly: 'ThreeYearly',
    FiveYearly: 'FiveYearly',
    Unknown: 'Unknown'
} as const;
export type ReportingPeriod = typeof ReportingPeriod[keyof typeof ReportingPeriod];


export function instanceOfReportingPeriod(value: any): boolean {
    for (const key in ReportingPeriod) {
        if (Object.prototype.hasOwnProperty.call(ReportingPeriod, key)) {
            if (ReportingPeriod[key as keyof typeof ReportingPeriod] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ReportingPeriodFromJSON(json: any): ReportingPeriod {
    return ReportingPeriodFromJSONTyped(json, false);
}

export function ReportingPeriodFromJSONTyped(json: any, ignoreDiscriminator: boolean): ReportingPeriod {
    return json as ReportingPeriod;
}

export function ReportingPeriodToJSON(value?: ReportingPeriod | null): any {
    return value as any;
}

export function ReportingPeriodToJSONTyped(value: any, ignoreDiscriminator: boolean): ReportingPeriod {
    return value as ReportingPeriod;
}